separate util class.
