+++ src/main/java/org/spin/grpc/service/MatchPOReceiptInvoiceServiceImplementation.java
@@ -27,14 +27,11 @@ import java.util.Properties;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.logging.Level;
 
-import org.adempiere.core.domains.models.I_AD_System;
 import org.adempiere.core.domains.models.I_C_BPartner;
 import org.adempiere.core.domains.models.I_C_Invoice;
-import org.adempiere.core.domains.models.I_C_InvoiceLine;
 import org.adempiere.core.domains.models.I_C_Order;
 import org.adempiere.core.domains.models.I_M_InOut;
 import org.adempiere.core.domains.models.I_M_Product;
-import org.compiere.model.MBPartner;
 import org.compiere.model.MLookupInfo;
 import org.compiere.model.MProduct;
 import org.compiere.model.MRole;
@@ -64,7 +61,6 @@ import org.spin.backend.grpc.form.match_po_receipt_invoice.Matched;
 import org.spin.backend.grpc.form.match_po_receipt_invoice.ProcessRequest;
 import org.spin.backend.grpc.form.match_po_receipt_invoice.ProcessResponse;
 import org.spin.backend.grpc.form.match_po_receipt_invoice.Product;
-import org.spin.backend.grpc.form.match_po_receipt_invoice.Vendor;
 import org.spin.backend.grpc.form.match_po_receipt_invoice.MatchPORReceiptInvoiceGrpc.MatchPORReceiptInvoiceImplBase;
 import org.spin.base.util.LookupUtil;
 import org.spin.base.util.RecordUtil;
@@ -306,40 +302,6 @@ public class MatchPOReceiptInvoiceServiceImplementation extends MatchPORReceiptI
 
 
 
-	public static Vendor.Builder convertVendor(int businessPartnerId) {
-		if (businessPartnerId <= 0) {
-			return Vendor.newBuilder();
-		}
-		MBPartner businessPartner = MBPartner.get(Env.getCtx(), businessPartnerId);
-		return convertVendor(businessPartner);
-	}
-	public static Vendor.Builder convertVendor(MBPartner businessPartner) {
-		Vendor.Builder builder = Vendor.newBuilder();
-		if (businessPartner == null || businessPartner.getC_BPartner_ID() <= 0) {
-			return builder;
-		}
-
-		builder.setId(businessPartner.getC_BPartner_ID())
-			.setUuid(
-				ValueUtil.validateNull(businessPartner.getUUID())
-			)
-			.setValue(
-				ValueUtil.validateNull(businessPartner.getValue())
-			)
-			.setTaxId(
-				ValueUtil.validateNull(businessPartner.getTaxID())
-			)
-			.setName(
-				ValueUtil.validateNull(businessPartner.getName())
-			)
-			.setDescription(
-				ValueUtil.validateNull(businessPartner.getDescription())
-			)
-		;
-
-		return builder;
-	}
-
 	@Override
 	public void listVendors(ListVendorsRequest request, StreamObserver<ListLookupItemsResponse> responseObserver) {
 		try {
@@ -497,192 +459,6 @@ public class MatchPOReceiptInvoiceServiceImplementation extends MatchPORReceiptI
 
 
 
-	public static String getDateColumn(int matchType) {
-		if (matchType == MatchType.INVOICE_VALUE) {
-			return "hdr.DateInvoiced";
-		}
-		else if (matchType == MatchType.PURCHASE_ORDER_VALUE) {
-			return "hdr.DateOrdered";
-		}
-		// Receipt
-		return "hdr.MovementDate";
-	}
-
-	public static String getQuantityColumn(int matchType) {
-		if (matchType == MatchType.INVOICE_VALUE) {
-			return "lin.QtyInvoiced";
-		}
-		else if (matchType == MatchType.PURCHASE_ORDER_VALUE) {
-			return "lin.QtyOrdered";
-		}
-		// Receipt
-		return "lin.MovementQty";
-	}
-
-	public static String getSQL(boolean isMatched, int matchTypeFrom, int matchTypeTo) {
-		String sql = "";
-		if (matchTypeFrom == MatchType.INVOICE_VALUE) {
-			sql = "SELECT lin.C_InvoiceLine_ID AS ID, lin.UUID AS UUID, "
-				+ " hdr.C_Invoice_ID AS Header_ID, hrd.Header_UUID, hdr.DateInvoiced AS Date,"
-				+ " hdr.C_Invoice_ID, hdr.DocumentNo, hdr.DateInvoiced, "
-				+ " bp.Name AS C_BPartner_Name, hdr.C_BPartner_ID, "
-				+ " lin.Line, lin.C_InvoiceLine_ID, "
-				+ " p.Name AS M_Product_Name, lin.M_Product_ID, "
-				+ " lin.QtyInvoiced, SUM(COALESCE(mi.Qty, 0)), org.Name, hdr.AD_Org_ID "
-				+ "FROM C_Invoice hdr"
-				+ " INNER JOIN AD_Org org ON (hdr.AD_Org_ID = org.AD_Org_ID)"
-				+ " INNER JOIN C_BPartner bp ON (hdr.C_BPartner_ID = bp.C_BPartner_ID)"
-				+ " INNER JOIN C_InvoiceLine lin ON (hdr.C_Invoice_ID = lin.C_Invoice_ID)"
-				+ " INNER JOIN M_Product p ON (lin.M_Product_ID = p.M_Product_ID)"
-				+ " INNER JOIN C_DocType dt ON (hdr.C_DocType_ID = dt.C_DocType_ID AND dt.DocBaseType IN ('API', 'APC'))"
-				+ " FULL JOIN M_MatchInv mi ON (lin.C_InvoiceLine_ID = mi.C_InvoiceLine_ID) "
-				+ "WHERE hdr.DocStatus IN ('CO','CL')"
-			;
-		}
-		else if (matchTypeFrom == MatchType.PURCHASE_ORDER_VALUE) {
-			String lineType = matchTypeTo == MatchType.RECEIPT_VALUE ? "M_InOutLine_ID" : "C_InvoiceLine_ID";
-
-			sql = "SELECT lin.C_OrderLine_ID AS ID, lin.UUID AS UUID, "
-				+ " hdr.C_Order_ID AS Header_ID, hrd.Header_UUID, hdr.DateOrdered AS Date,"
-				+ " hdr.C_Order_ID, hdr.DocumentNo, hdr.DateOrdered, "
-				+ " bp.Name AS C_BPartner_Name, hdr.C_BPartner_ID, "
-				+ " lin.Line, lin.C_OrderLine_ID, "
-				+ " p.Name AS M_Product_Name, lin.M_Product_ID, "
-				+ " lin.QtyOrdered, SUM(COALESCE(mo.Qty, 0)), org.Name, hdr.AD_Org_ID "
-				+ "FROM C_Order hdr"
-				+ " INNER JOIN AD_Org org ON (hdr.AD_Org_ID = org.AD_Org_ID)"
-				+ " INNER JOIN C_BPartner bp ON (hdr.C_BPartner_ID = bp.C_BPartner_ID)"
-				+ " INNER JOIN C_OrderLine lin ON (hdr.C_Order_ID = lin.C_Order_ID)"
-				+ " INNER JOIN M_Product p ON (lin.M_Product_ID = p.M_Product_ID)"
-				+ " INNER JOIN C_DocType dt ON (hdr.C_DocType_ID = dt.C_DocType_ID AND dt.DocBaseType = 'POO')"
-				+ " FULL JOIN M_MatchPO mo ON (lin.C_OrderLine_ID = mo.C_OrderLine_ID) "
-				+ " WHERE "
-			;
-			
-			if (isMatched) {
-				sql += " mo." + lineType + " IS NOT NULL "; 
-			} else {
- 				sql += " ( mo." + lineType + " IS NULL OR "
-					+ " (lin.QtyOrdered <> (SELECT sum(mo1.Qty) AS Qty"
-					+ " FROM m_matchpo mo1 WHERE "
-					+ " mo1.C_ORDERLINE_ID = lin.C_ORDERLINE_ID AND "
-					+ " hdr.C_ORDER_ID = lin.C_ORDER_ID AND "
-					+ " mo1." + lineType
-					+ " IS NOT NULL group by mo1.C_ORDERLINE_ID))) "
-				;
-			}
-			sql += " AND hdr.DocStatus IN ('CO', 'CL') ";
-		}
-		// Receipt
-		else {
-			sql = "SELECT lin.M_InOutLine_ID AS ID, lin.UUID AS UUID, "
-				+ " hdr.M_InOut_ID AS Header_ID, hrd.Header_UUID, hdr.MovementDate AS Date,"
-				+ " hdr.M_InOut_ID, hdr.DocumentNo, hdr.MovementDate, "
-				+ " bp.Name AS C_BPartner_Name, hdr.C_BPartner_ID, "
-				+ " lin.Line, lin.M_InOutLine_ID, "
-				+ " p.Name AS M_Product_Name, lin.M_Product_ID, "
-				+ " lin.MovementQty, SUM(COALESCE(m.Qty, 0)), org.Name, hdr.AD_Org_ID "
-				+ "FROM M_InOut hdr"
-				+ " INNER JOIN AD_Org org ON (hdr.AD_Org_ID = org.AD_Org_ID)"
-				+ " INNER JOIN C_BPartner bp ON (hdr.C_BPartner_ID = bp.C_BPartner_ID)"
-				+ " INNER JOIN M_InOutLine lin ON (hdr.M_InOut_ID = lin.M_InOut_ID)"
-				+ " INNER JOIN M_Product p ON (lin.M_Product_ID = p.M_Product_ID)"
-				+ " INNER JOIN C_DocType dt ON (hdr.C_DocType_ID = dt.C_DocType_ID AND dt.DocBaseType IN ('MMR', 'MMS'))"
-				+ " FULL JOIN "
-				+ (matchTypeFrom == MatchType.PURCHASE_ORDER_VALUE ? "M_MatchPO" : "M_MatchInv")
-				+ " m ON (lin.M_InOutLine_ID = m.M_InOutLine_ID) "
-				+ "WHERE hdr.DocStatus IN ('CO', 'CL') and dt.issotrx = 'N' "
-			;
-		}
-		return sql;
-	}
-
-	public static String getGroupBy(boolean isMatched, int matchType) {
-		if (matchType == MatchType.INVOICE_VALUE) {
-			return " GROUP BY hdr.C_Invoice_ID,hdr.DocumentNo,hdr.DateInvoiced,bp.Name,hdr.C_BPartner_ID,"
-				+ " lin.Line,lin.C_InvoiceLine_ID,p.Name,lin.M_Product_ID,lin.QtyInvoiced, org.Name, hdr.AD_Org_ID "
-				+ "HAVING "
-				+ (isMatched ? "0" : "lin.QtyInvoiced")
-				+ "<>SUM(COALESCE(mi.Qty,0))"
-			;
-		}
-		else if (matchType == MatchType.PURCHASE_ORDER_VALUE) {
-			return " GROUP BY hdr.C_Order_ID,hdr.DocumentNo,hdr.DateOrdered,bp.Name,hdr.C_BPartner_ID,"
-				+ " lin.Line,lin.C_OrderLine_ID,p.Name,lin.M_Product_ID,lin.QtyOrdered, org.Name, hdr.AD_Org_ID "
-				+ "HAVING "
-				+ (isMatched ? "0" : "lin.QtyOrdered")
-				+ "<>SUM(COALESCE(mo.Qty,0))";
-		}
-		// Receipt
-		return " GROUP BY hdr.M_InOut_ID,hdr.DocumentNo,hdr.MovementDate,bp.Name,hdr.C_BPartner_ID,"
-			+ " lin.Line,lin.M_InOutLine_ID,p.Name,lin.M_Product_ID,lin.MovementQty, org.Name, hdr.AD_Org_ID "
-			+ "HAVING "
-			+ (isMatched ? "0" : "lin.MovementQty")
-			+ "<>SUM(COALESCE(m.Qty,0))"
-		;
-	}
-
-	public static Matched.Builder convertMatched(ResultSet resultSet) throws SQLException {
-		Matched.Builder builder = Matched.newBuilder();
-		if (resultSet == null) {
-			return builder;
-		}
-
-		builder.setId(
-				resultSet.getInt("ID")
-			)
-			.setUuid(
-				ValueUtil.validateNull(
-					resultSet.getString(I_AD_System.COLUMNNAME_UUID)
-				)
-			)
-			.setHeaderId(
-				resultSet.getInt("Header_ID")
-			)
-			.setHeaderUuid(
-				ValueUtil.validateNull(
-					resultSet.getString("Header_UUID")
-				)
-			)
-			.setDate(
-				ValueUtil.getLongFromTimestamp(
-					resultSet.getTimestamp("Date")
-				)
-			)
-			.setDocumentNo(
-				ValueUtil.validateNull(
-					resultSet.getString(I_C_Invoice.COLUMNNAME_DocumentNo)
-				)
-			)
-			.setLineNo(
-				resultSet.getInt(I_C_InvoiceLine.COLUMNNAME_Line)
-			)
-			.setQuantity(
-				ValueUtil.getDecimalFromBigDecimal(
-					resultSet.getBigDecimal("Quantity")
-				)
-			)
-			.setProductId(
-				resultSet.getInt(I_C_InvoiceLine.COLUMNNAME_M_Product_ID)
-			)
-			.setProductName(
-				ValueUtil.validateNull(
-					resultSet.getString("M_Product_Name")
-				)
-			)
-			.setVendorId(
-				resultSet.getInt(I_C_Invoice.COLUMNNAME_C_BPartner_ID)
-			)
-			.setVendorName(
-				ValueUtil.validateNull(
-					resultSet.getString("C_BPartner_Name")
-				)
-			)
-		;
-
-		return builder;
-	}
-
 	@Override
 	public void listMatchedFrom(ListMatchedFromRequest request, StreamObserver<ListMatchedFromResponse> responseObserver) {
 		try {
@@ -708,10 +484,9 @@ public class MatchPOReceiptInvoiceServiceImplementation extends MatchPORReceiptI
 		int matchFromType = request.getMatchFromTypeValue();
 		int matchToType = request.getMatchToTypeValue();
 
-		final String dateColumn = getDateColumn(matchFromType);
-		// final String quantityColumn = getQuantityColumn(matchFromType);
-		final String sql = getSQL(isMatched, matchFromType, matchToType);
-		final String groupBy = getGroupBy(isMatched, matchFromType);
+		final String dateColumn = org.spin.form.match_po_receipt_invoice.Util.getDateColumn(matchFromType);
+		final String sql = org.spin.form.match_po_receipt_invoice.Util.getSQL(isMatched, matchFromType, matchToType);
+		final String groupBy = org.spin.form.match_po_receipt_invoice.Util.getGroupBy(isMatched, matchFromType);
 
 		String whereClause = "";
 		if (request.getProductId() > 0) {
@@ -753,7 +528,7 @@ public class MatchPOReceiptInvoiceServiceImplementation extends MatchPORReceiptI
 			while (rs.next()) {
 				recordCount++;
 
-				Matched.Builder matchedBuilder = convertMatched(rs);
+				Matched.Builder matchedBuilder = org.spin.form.match_po_receipt_invoice.Util.convertMatched(rs);
 				builderList.addRecords(matchedBuilder);
 			}
 		}
@@ -799,10 +574,9 @@ public class MatchPOReceiptInvoiceServiceImplementation extends MatchPORReceiptI
 		int matchToType = request.getMatchToTypeValue();
 		int matchFromSelectedId = request.getMatchFromSelectedId();
 
-		final String dateColumn = getDateColumn(matchFromType);
-		// final String quantityColumn = getQuantityColumn(matchFromType);
-		final String sql = getSQL(isMatched, matchFromType, matchToType);
-		final String groupBy = getGroupBy(isMatched, matchFromType);
+		final String dateColumn = org.spin.form.match_po_receipt_invoice.Util.getDateColumn(matchFromType);
+		final String sql = org.spin.form.match_po_receipt_invoice.Util.getSQL(isMatched, matchFromType, matchToType);
+		final String groupBy = org.spin.form.match_po_receipt_invoice.Util.getGroupBy(isMatched, matchFromType);
 
 		String whereClause = "";
 		if (request.getProductId() > 0) {
@@ -826,8 +600,8 @@ public class MatchPOReceiptInvoiceServiceImplementation extends MatchPORReceiptI
 		}
 
 		if (request.getIsSameQuantity()) {
-			final String quantityColumn = getQuantityColumn(matchFromType);
-			Matched.Builder matchedFromSelected = getMatchedSelectedFrom(matchFromSelectedId, isMatched, matchFromType, matchToType);
+			final String quantityColumn = org.spin.form.match_po_receipt_invoice.Util.getQuantityColumn(matchFromType);
+			Matched.Builder matchedFromSelected = org.spin.form.match_po_receipt_invoice.Util.getMatchedSelectedFrom(matchFromSelectedId, isMatched, matchFromType, matchToType);
 			BigDecimal quantity = ValueUtil.getBigDecimalFromDecimal(
 				matchedFromSelected.getQuantity()
 			);
@@ -853,7 +627,7 @@ public class MatchPOReceiptInvoiceServiceImplementation extends MatchPORReceiptI
 			while (rs.next()) {
 				recordCount++;
 
-				Matched.Builder matchedBuilder = convertMatched(rs);
+				Matched.Builder matchedBuilder = org.spin.form.match_po_receipt_invoice.Util.convertMatched(rs);
 				builderList.addRecords(matchedBuilder);
 			}
 		}
@@ -871,47 +645,7 @@ public class MatchPOReceiptInvoiceServiceImplementation extends MatchPORReceiptI
 		return builderList;
 	}
 
-	public Matched.Builder getMatchedSelectedFrom(int matchFromSelectedId, boolean isMatched, int matchTypeFrom, int matchTypeTo) {
-		Matched.Builder builder = Matched.newBuilder();
-		if (matchFromSelectedId <= 0) {
-			return builder;
-		}
 
-		// Receipt
-		String whereClause = " AND lin.M_InOutLine_ID = ";
-		if (matchTypeFrom == MatchType.INVOICE_VALUE) {
-			whereClause = " AND lin.C_InvoiceLine_ID = ";
-		}
-		else if (matchTypeFrom == MatchType.PURCHASE_ORDER_VALUE) {
-			whereClause = " AND lin.C_OrderLine_ID = ";
-		}
-		whereClause += matchFromSelectedId;
-
-		final String sql = getSQL(isMatched, matchTypeFrom, matchTypeTo);
-		final String groupBy = getGroupBy(isMatched, matchTypeFrom);
-
-		final String sqlWithAccess = MRole.getDefault().addAccessSQL(
-			sql + whereClause,
-			"hdr",
-			MRole.SQL_FULLYQUALIFIED,
-			MRole.SQL_RO
-		) + groupBy;
-
-		PreparedStatement pstmt = null;
-		ResultSet rs = null;
-		try {
-			pstmt = DB.prepareStatement(sqlWithAccess, null);
-			rs = pstmt.executeQuery();
-
-			if (rs.next()) {
-				builder = convertMatched(rs);
-			}
-		} catch (Exception e) {
-			// TODO: handle exception
-		}
-
-		return builder;
-	}
 
 	@Override
 	public void process(ProcessRequest request, StreamObserver<ProcessResponse> responseObserver) {
